Network/requirements.txt
fire==0.4.0
PyYAML==6.0
six==1.16.0
termcolor==2.0.1
